refactor: modify dependent code by renamed search type
